Fix component structure and exports
