+++ src/components/RestaurantListing/RestaurantListing.tsx
@@ -1,0 +1,71 @@
+import React, { useState } from 'react';
+import { Card, CardHeader, CardTitle, CardContent } from '@/components/ui/card';
+import { Star, Search } from 'lucide-react';
+
+// Filter types
+interface Filters {
+  rating: number;
+  priceRange: string;
+  cuisine: string;
+}
+
+// Restaurant interface
+interface Restaurant {
+  name: string;
+  address: string;
+  rating: number;
+  priceRange: string;
+  cuisine: string;
+  image: string;
+  review: {
+    text: string;
+    author: string;
+    rating: number;
+  };
+  hours: string;
+  phone: string;
+  location: {
+    lat: number;
+    lng: number;
+  };
+}
+
+const RestaurantListing = () => {
+  const allRestaurants = [
+    {
+      name: "Boho Café",
+      address: "10 Rue EL Yamama, Rabat 10000, Morocco",
+      image: "/api/placeholder/800/400?text=Bright+Modern+Café",
+      rating: 4.7,
+      priceRange: "$$",
+      cuisine: "Café",
+      review: {
+        text: "Loved this brunch place! Lots of healthy options.",
+        author: "kapil gupta",
+        rating: 5
+      },
+      hours: "9:00 AM – 9:00 PM",
+      phone: "0690-036170",
+      location: { lat: 34.0151034, lng: -6.8369946 }
+    }
+  ];
+
+  const [filters, setFilters] = useState<Filters>({
+    rating: 0,
+    priceRange: 'all',
+    cuisine: 'all'
+  });
+
+  const [searchQuery, setSearchQuery] = useState('');
+  
+  return (
+    <div className="max-w-4xl mx-auto p-4 space-y-6">
+      <h1 className="text-3xl font-bold text-center mb-8">
+        Top Restaurants in Rabat
+      </h1>
+      <div>Restaurant Listing Component</div>
+    </div>
+  );
+};
+
+export default RestaurantListing;
