+++ src/components/RestaurantListing/index.tsx
@@ -1,39 +1,1 @@
-import React, { useState } from 'react';
-import { Card, CardHeader, CardTitle, CardContent } from '@/components/ui/card';
-import { Star, Search } from 'lucide-react';
-import { Restaurant, Filters } from './types';
-
-const RestaurantListing = () => {
-  const allRestaurants = [
-    {
-      name: "Boho Café",
-      address: "10 Rue EL Yamama, Rabat 10000, Morocco",
-      image: "/api/placeholder/800/400?text=Bright+Modern+Café",
-      rating: 4.7,
-      priceRange: "$$",
-      cuisine: "Café",
-      review: {
-        text: "Loved this brunch place! Lots of healthy options.",
-        author: "kapil gupta",
-        rating: 5
-      },
-      hours: "9:00 AM – 9:00 PM",
-      phone: "0690-036170",
-      location: { lat: 34.0151034, lng: -6.8369946 }
-    }
-  ];
-
-  const [filters, setFilters] = useState<Filters>({
-    rating: 0,
-    priceRange: 'all',
-    cuisine: 'all'
-  });
-
-  const [searchQuery, setSearchQuery] = useState('');
-  
-  return (
-    <div>Restaurant Listing Component</div>
-  );
-};
-
-export default RestaurantListing;
+export { default as RestaurantListing } from './RestaurantListing';
